% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa_codon_translation.R
\name{aa_codon_translation}
\alias{aa_codon_translation}
\title{amino acids to codons translation}
\usage{
aa_codon_translation(aa_seq, start = 1)
}
\arguments{
\item{aa_seq}{original sequence of amino acids}

\item{start}{first position of the input sequence}
}
\value{
codons sequence corresponding to the aa original sequence
}
\description{
amino acids to codons translation
}
