% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_five.R
\name{aa_count}
\alias{aa_count}
\title{Occurence of amino acids in a peptide}
\usage{
aa_count(peptide_sequence)
}
\arguments{
\item{peptide_sequence}{Supply a peptide, it will count every character as a string and display it in a barplot in a plot.}
}
\value{
occurrence The count of every single amino acid character
}
\description{
Occurence of amino acids in a peptide
}
