% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join.R
\name{AAseq}
\alias{AAseq}
\title{Concatenate codons}
\usage{
AAseq(codons)
}
\arguments{
\item{codons}{3-letters}
}
\value{
the output of the function is the single-letter codon from an input
3-letter codon.
}
\description{
Concatenate codons
}
\examples{
AAseq("UUU")
}
