% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_three.R
\name{codon_split}
\alias{codon_split}
\title{Function number three}
\usage{
codon_split(seq_DNA, start = 1)
}
\arguments{
\item{seq_DNA}{DNA sequence}

\item{codon_split}{splits DNA into codons}
}
\value{
it returns codons from DNA sequence
}
\description{
Function number three
}
\examples{
codon_split("CUCCUACUG")
}
