% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codon table.R, R/data.R
\docType{data}
\name{codon_table}
\alias{codon_table}
\title{Codon_table}
\format{
An object of class \code{character} of length 64.


}
\source{
\url{https://www.ncbi.nlm.nih.gov/Taxonomy/Utils/wprintgc.cgi?chapter=tgencodes#SG1}

\url{https://www.ncbi.nlm.nih.gov/Taxonomy/Utils/wprintgc.cgi?chapter=tgencodes#SG1}
}
\usage{
codon_table

codon_table
}
\arguments{
\item{codon_table}{shows to which amino acid each codon will be translated to .}
}
\value{

}
\description{
Description Codon table
}
\keyword{datasets}
