% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dna.R
\name{create_dna}
\alias{create_dna}
\title{Title Create random DNA with given size}
\usage{
create_dna(size)
}
\arguments{
\item{size}{integer value that represent the length of the DNA sequence to be randomlygenerated .}
}
\value{

}
\description{
Title Create random DNA with given size
}
