% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DNA_to_RNA.R
\name{DNA_to_RNA}
\alias{DNA_to_RNA}
\title{Primitive function to translate a DNA sequence into a RNA sequence}
\usage{
DNA_to_RNA(dna_seq)
}
\arguments{
\item{dna_seq}{DNA sequence}
}
\value{
RNA sequence
}
\description{
Primitive function to translate a DNA sequence into a RNA sequence
}
\examples{
DNA_to_RNA("ATCGTTG")
}
