% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_4.R
\name{func_4}
\alias{func_4}
\title{Example function}
\usage{
func_4(codons)
}
\arguments{
\item{codons}{A character vector of RNA codons (e.g., c('AUG', 'UUU', 'GGC')).}
}
\value{
A string representing the translated amino acid sequence.
}
\description{
Example function
}
