% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_dna_string.R
\name{generate_dna_string}
\alias{generate_dna_string}
\title{Generate random DNA sequence}
\usage{
generate_dna_string(length)
}
\arguments{
\item{length}{The desired length of the sequence}
}
\value{
string that represents DNA sequence
}
\description{
Generate random DNA sequence
}
\examples{
generate_dna_string(10)
generate_dna_string(25)
}
