% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_sequence.R
\name{generated_sequence}
\alias{generated_sequence}
\title{Random sequence}
\usage{
generated_sequence(number_nucleotides)
}
\arguments{
\item{number_nucleotides}{}
}
\value{
the function will return sequence corresponding with the number of nucleotides
}
\description{
Random sequence
}
