% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_reading_frame.R
\name{get_reading_frame}
\alias{get_reading_frame}
\title{Reading frame}
\usage{
get_reading_frame(DNA_seq, start = 1)
}
\arguments{
\item{DNA_seq}{String  DNA sequence to be translated with the translate function  .}

\item{start}{Integer value startingpoint of the reading frame .}
}
\value{
returns the codons character vector of three base pair to be subsequently translated to amino acids
}
\description{
Reading frame
}
\examples{
get_reading_frame("ATGCGC", start =  1))
will return [ "ATG"  , "CGC"]

}
