% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_aa_frequence.R
\name{plot_aa_frequence}
\alias{plot_aa_frequence}
\title{Frequency of amino acids}
\usage{
plot_aa_frequence(name_me2)
}
\arguments{
\item{name_me2}{}
}
\value{
name_me4 a plot with the amino acid frequencies
}
\description{
Frequency of amino acids
}
