% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_aa_frequence.R
\name{plot_aa_frequency}
\alias{plot_aa_frequency}
\title{Frequency of amino acids}
\usage{
plot_aa_frequency(amino_acid_sequence)
}
\arguments{
\item{amino_acid_sequence}{A character vector representing the amino acid sequence.}
}
\value{
amino_acid_frequency_plot A plot with the amino acid counts of the given amino acid sequence.
}
\description{
This function calculates and plots the frequency of amino acids in a given amino acid sequence.
}
