% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_frequencies.R
\name{plot_frequencies}
\alias{plot_frequencies}
\title{Count Unique Characters and Plot Frequency}
\usage{
plot_frequencies(sequence_str)
}
\arguments{
\item{sequence_str}{A sequence string to be analyzed for unique nucleotide frequencies.}
}
\value{
A ggplot2 bar plot showing the frequency of each unique nucleotide in the input string.
}
\description{
This function counts the frequency of unique nucleotide in a string and generates a bar plot
showing the counts of each unique character.
}
\examples{
plot_frequencies("AGTCAGTC")
# Returns: A ggplot bar plot with counts for each unique nucleotide

}
