% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sequence_counts.R
\name{plot_sequence_counts}
\alias{plot_sequence_counts}
\title{Plots sequence counts}
\usage{
plot_sequence_counts(sequence_string)
}
\arguments{
\item{sequence}{input string of characters to be analyse}
}
\value{
ggplot 2 plot of sequence counts
}
\description{
Plots sequence counts
}
\examples{
plot_sequence_counts("ATGCGTAC")
}
