% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_dna.R
\name{random_dna_seq}
\alias{random_dna_seq}
\title{Title: Random DNA sequence}
\usage{
random_dna_seq(length)
}
\arguments{
\item{length}{Length of desired random DNA sequence}
}
\value{
The function allows the user to create random DNA sequences with desired length
}
\description{
Title: Random DNA sequence
}
\examples{
random_dna_seq(20)
}
