% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DNA_2_RNA.R
\name{RNA}
\alias{RNA}
\title{DNA_2_RNA}
\usage{
RNA(DNA)
}
\arguments{
\item{DNA}{a DNA string}
}
\value{
a RNA_string
}
\description{
DNA_2_RNA
}
\examples{
# Convert a simple DNA sequence to RNA
RNA("ATGC")

# You can also store the result in a variable
my_RNA <- RNA("TTGCA")
my_RNA

# Expected output:
# "UUGCA"
}
