% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_to_codons.R
\name{split_to_codons}
\alias{split_to_codons}
\title{Split a String into Codons}
\usage{
split_to_codons(sequence, start = 1)
}
\arguments{
\item{sequence}{A sequence (string) to be split into codons.}

\item{start}{An integer indicating the starting position to begin forming codons. Default is 1.}
}
\value{
A character vector of codons (3-character substrings).
}
\description{
This function takes a character string and splits it into triplets of characters (codons).
By default, it starts from the first character, but a custom starting position can be provided.
}
\examples{
split_to_codons("ATGCGTACG", start = 1)
# Returns: "ATG" "CGT" "ACG"

split_to_codons("GTCAGTCTTAG", start = 2)
# Returns: "TCA" "GTC" "TAG"

}
