% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcribe.R
\name{transcribe}
\alias{transcribe}
\title{Convert a sense (coding) DNA sequence into mRNA}
\usage{
transcribe(dna_seq)
}
\arguments{
\item{dna_seq, }{character vector}
}
\value{
Returns an mRNA sequence (character vector).
}
\description{
Convert a sense (coding) DNA sequence into mRNA
}
\examples{
transcribe("ACTG")
}
