% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcription.R
\name{transcripted_sequence}
\alias{transcripted_sequence}
\title{Transcription}
\usage{
transcripted_sequence(DNA_sequence)
}
\arguments{
\item{DNA_sequence}{a DNA sequence to be transcribed}
}
\value{
Returns a RNA sequence of the inserted DNA sequence
}
\description{
Transcription
}
