% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translation.R
\name{translate}
\alias{translate}
\title{Concatenate codons}
\usage{
translate(codons)
}
\arguments{
\item{codons}{a set of RNA codons, can be a vector or 3-letter string}
}
\value{
the translated protein sequence from the RNA codons
}
\description{
Concatenate codons
}
\examples{
translate("UUU")
translate(c("UUU","UCU"))

}
