% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_five.R
\name{aa_count}
\alias{aa_count}
\title{Occurence of amino acids in a peptide}
\usage{
aa_count(amino_acid)
}
\arguments{
\item{aa_count}{}
}
\value{

}
\description{
Occurence of amino acids in a peptide
}
