% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{codon_table}
\alias{codon_table}
\title{Title : codon_table}
\format{
An object of class \code{character} of length 64.
}
\usage{
codon_table
}
\description{
Description: Codon table from class in week 8
}
\keyword{datasets}
