% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_aa_frequence.R
\name{plot_aa_frequence}
\alias{plot_aa_frequence}
\title{Frequency of amino acids}
\usage{
plot_aa_frequence(amino_acid_sequence)
}
\arguments{
\item{amino_acid_sequence}{}
}
\value{
amino_acid_frequency_plot a plot with the amino acid counts of a given amino acid sequence
}
\description{
Frequency of amino acids
}
